Fix bill value to be Long instead of Int

+++ app/src/main/java/au/com/example/tipcalculator/MainViewModel.kt
@@ -14,10 +14,10 @@ class MainViewModel : ViewModel() {
     val tipValue: LiveData<String>
         get() = _tipValue
 
-    private var billTotal = 0
+    private var billTotal = 0L
     private var percentageTip = 0
 
-    fun onTotalChanged(billTotal: Int) {
+    fun onTotalChanged(billTotal: Long) {
         this.billTotal = billTotal
         calculateTip()
     }
